# src/infra/controllers/http/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FindBar Assessment API",
    "version": "1.0.0"
  },
  "paths": {
    "/products": {
      "post": {
        "summary": "Upload product JSON files",
        "description": "Accepts `.json` files containing product data. Files are stored and processed server-side.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "products": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": ["products"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data uploaded successfully."
          },
          "400": {
            "description": "Invalid file type. Only `.json` files are allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error or Multer upload error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Search for products",
        "description": "Performs a search for products based on a query string.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search query string",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "score": {
                        "type": "number"
                      },
                      "title": {
                        "type": "string"
                      }
                    },
                    "required": ["score", "title"]
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}